requirements.txt
opencv-python==4.6.0.66
tifffile==2022.10.10
pandas==1.3.5
scipy==1.10.1
yacs==0.1.8
lap==0.4.0
matplotlib==3.6.0
torch-geometric==2.5.0
filterpy==1.4.5
torch-scatter
motmetrics==1.4.0
